# Addon-Test: Make sure that only one global portable story config is ever loaded