--- a/code/addons/test/src/vitest-plugin/setup-file.ts
+++ b/code/addons/test/src/vitest-plugin/setup-file.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
 /* eslint-disable no-underscore-dangle */
-import { afterEach, beforeAll, vi } from 'vitest';
+import { afterEach, vi } from 'vitest';
 import type { RunnerTask } from 'vitest';
 
 import { Channel } from 'storybook/internal/channels';
@@ -35,10 +35,11 @@ export const modifyErrorMessage = ({ task }: { task: Task }) => {
   }
 };
 
-beforeAll(() => {
-  if (globalThis.csf4Preview) {
-    return globalThis.csf4Preview.composed.beforeAll();
-  }
-});
+// Enable this in 9.0
+// beforeAll(() => {
+//   if (globalThis.globalProjectAnnotations) {
+//     return globalThis.globalProjectAnnotations.beforeAll();
+//   }
+// });
 
 afterEach(modifyErrorMessage);