--- a/code/core/src/preview-api/modules/store/csf/portable-stories.ts
+++ b/code/core/src/preview-api/modules/store/csf/portable-stories.ts
@@ -41,8 +41,6 @@ declare global {
   var globalProjectAnnotations: NormalizedProjectAnnotations<any>;
   // eslint-disable-next-line no-var
   var defaultProjectAnnotations: ProjectAnnotations<any>;
-  // eslint-disable-next-line no-var
-  var csf4Preview: Preview<any>;
 }
 
 export function setDefaultProjectAnnotations<TRenderer extends Renderer = Renderer>(