<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  80.5 MB | 80.5 MB | 29 kB | **-1.4** | 0% |
| initSize |  80.5 MB | 80.5 MB | 29 kB | **-1.4** | 0% |
| diffSize |  97 B | 97 B | 0 B | - | 0% |
| buildSize |  7.31 MB | 7.31 MB | 263 B | **8.29** | 0% |
| buildSbAddonsSize |  1.9 MB | 1.9 MB | 129 B | **-2.8** | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.88 MB | 1.88 MB | 0 B | **4.36** | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.97 MB | 3.97 MB | 129 B | **6.26** | 0% |
| buildPreviewSize |  3.34 MB | 3.34 MB | 134 B | **26.73** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  15.9s | 12.5s | -3s -424ms | -0.28 | -27.4% |
| generateTime |  20.8s | 19s | -1s -780ms | -0.66 | -9.3% |
| initTime |  4.7s | 4.4s | -263ms | -0.64 | -5.9% |
| buildTime |  10.5s | 8.6s | -1s -918ms | -0.89 | -22.2% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.1s | 4.8s | -293ms | -1.13 | -6.1% |
| devManagerResponsive |  3.8s | 3.5s | -311ms | **-1.28** | 🔰-8.7% |
| devManagerHeaderVisible |  871ms | 754ms | -117ms | -0.4 | -15.5% |
| devManagerIndexVisible |  907ms | 764ms | -143ms | -0.52 | -18.7% |
| devStoryVisibleUncached |  4.1s | 3.9s | -186ms | 0.08 | -4.7% |
| devStoryVisible |  907ms | 825ms | -82ms | -0.27 | -9.9% |
| devAutodocsVisible |  841ms | 758ms | -83ms | -0.31 | -10.9% |
| devMDXVisible |  823ms | 701ms | -122ms | -0.91 | -17.4% |
| buildManagerHeaderVisible |  683ms | 581ms | -102ms | **-1.85** | 🔰-17.6% |
| buildManagerIndexVisible |  695ms | 622ms | -73ms | **-1.73** | 🔰-11.7% |
| buildStoryVisible |  667ms | 568ms | -99ms | **-1.74** | 🔰-17.4% |
| buildAutodocsVisible |  620ms | 490ms | -130ms | -1.18 | -26.5% |
| buildMDXVisible |  571ms | 508ms | -63ms | **-1.47** | 🔰-12.4% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Based on the provided information, I'll create a concise summary of the pull request changes:

Enhanced Vitest integration in Storybook's addon-test by adding automatic detection and execution of beforeAll lifecycle methods.

- Added caching mechanism in `csf-factories.ts` for composed project annotations to improve performance
- Modified `test-utils.ts` to use `composeConfigs` for handling story previews with CSF3/CSF4 compatibility
- Introduced `globalThis.csf4Preview` in `setup-file.ts` to store and access preview instance globally
- Updated `portable-stories.ts` to properly compose default and global annotations for project configuration



<!-- /greptile_comment -->